+++ src/main.ts
@@ -7,6 +7,8 @@ import { AppModuleCore } from './core/app.module.core';
 import { SwaggerModuleCore } from './core/swagger.module.core';
 import { getWAHAVersion, VERSION, WAHAVersion } from './version';
 import { WAHA_WEBHOOKS } from './structures/webhooks.dto';
+import { LogLevel } from '@nestjs/common';
+import { getLogLevels } from './helpers';
 
 async function loadModules(): Promise<
   [typeof AppModuleCore, typeof SwaggerModuleCore]
@@ -30,10 +32,7 @@ async function loadModules(): Promise<
 async function bootstrap() {
   const [AppModule, SwaggerModule] = await loadModules();
   const app = await NestFactory.create(AppModule, {
-    logger:
-      process.env.DEBUG != undefined
-        ? ['log', 'debug', 'error', 'verbose', 'warn']
-        : ['log', 'error', 'warn'],
+    logger: getLogLevels(),
   });
 
   app.enableShutdownHooks();