+++ src/structures/enums.dto.ts
@@ -1,6 +1,7 @@
 export const SECOND = 1000;
 
 export enum WAHAEvents {
+  SESSION_STATUS = 'session.status',
   MESSAGE = 'message',
   MESSAGE_ANY = 'message.any',
   MESSAGE_ACK = 'message.ack',